Restructured claim command code
At Ren's request.
